successfully use context to pass default props down to child elements
