+++ src/Draggable/Draggable.component.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import styled from 'styled-components';
-import { useParentProps } from './store';
+import DraggableContext from './DraggableContext';
 
 type Props = {
   /**
@@ -61,8 +61,11 @@ const SDraggable = styled.div`
 `;
 
 export const Draggable: React.FunctionComponent<Props> = props => {
-  useParentProps(props);
-  return <SDraggable {...props}>{props.children}</SDraggable>;
+  return (
+    <DraggableContext.Provider value={{ parentProps: props }}>
+      <SDraggable {...props}>{props.children}</SDraggable>
+    </DraggableContext.Provider>
+  );
 };
 Draggable.defaultProps = {
   color: 'lightGray',
