+++ src/Draggable/DraggableContext.ts
@@ -1,8 +1,37 @@
 import * as React from 'react';
 
-const DraggableContext = React.createContext({
-  props: {},
-  updateProps: () => {},
+export const DraggableContext = React.createContext({
+  parentProps: {},
 });
+
+export const useMergeWithParentProps: React.FunctionComponent<any> = (
+  localProps,
+  propsToMerge,
+) => {
+  const parentProps = React.useContext(DraggableContext).parentProps;
+  const [newProps, setNewProps] = React.useState(localProps);
+
+  function mergeProps() {
+    const mergedProps = { ...parentProps, ...localProps };
+    propsToMerge.map((p: any) => {
+      mergedProps[p.key] =
+        localProps[p.key] || parentProps[p.key] || p.defaultVal;
+    });
+    return mergedProps;
+  }
+
+  React.useEffect(() => {
+    if (
+      parentProps !== null &&
+      JSON.stringify(localProps) !== JSON.stringify(parentProps)
+    ) {
+      const mergedProps = mergeProps();
+      if (JSON.stringify(newProps) !== JSON.stringify(mergedProps)) {
+        setNewProps(mergedProps);
+      }
+    }
+  }, [localProps, parentProps]);
+  return newProps;
+};
 
 export default DraggableContext;
